requirements.txt
black
redis
torch
openai
pytest
pymongo
pydantic
# psycopg2
pytest-mock
transformers
sentencepiece
pytest-asyncio
psycopg2-binary
pylint[spelling]
python-telegram-bot

# Async Database
sqlalchemy[asyncio]
alembic
asyncpg  
aiosqlite  

# Optional (Web Dashboard)
fastapi
uvicorn[standard]
jinja2
